requirements.txt
allure-pytest==2.8.18
pytest==6.0.1
requests==2.28.1
pytest-xdist==2.1.0
selenium==4.5.0
webdriver_manager==2.5.0
keyboard~=0.13.5